--- aries-core/src/js/components/core/NavigationMenu/index.ts
@@ -1,2 +1,0 @@
-export * from './NavigationMenu';
-export * from './NavItem';
