+++ aries-core/src/js/components/core/NavigationPane/index.ts
@@ -1,0 +1,2 @@
+export * from './NavigationPane';
+export * from './NavItem';
