+++ aries-core/src/js/components/core/index.js
@@ -1,5 +1,5 @@
 export * from './Identifier';
-export * from './NavigationMenu/index.ts';
+export * from './NavigationPane/index.ts';
 export * from './Popover';
 export * from './RangeSelector';
 export * from './Tile';
